requirements.txt
discord.py>=2.3.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
asyncio